# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/departments": {
      "get": {
        "operationId": "DepartmentsController_getAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department list returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "List departments",
        "tags": [
          "departments"
        ]
      },
      "post": {
        "operationId": "DepartmentsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Create a department",
        "tags": [
          "departments"
        ]
      }
    },
    "/api/departments/{id}": {
      "patch": {
        "operationId": "DepartmentsController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "D101",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Update a department",
        "tags": [
          "departments"
        ]
      },
      "delete": {
        "operationId": "DepartmentsController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "D101",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Delete a department",
        "tags": [
          "departments"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResourceX Demo Backend",
    "description": "Demo-exact NestJS backend for the ResourceX frontend. All authorization is role-based through request headers. Use x-user-role on every protected route and x-user-id when actor-specific scoping is required.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Required RBAC header. Example values: Requestor, Dept Head, Registrar, Staff, System Admin."
      },
      "user-id-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Optional acting user id header for own-profile and department-scoped behavior. Example: U100."
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Selected resources must match request quantity."
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "DepartmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "D1"
          },
          "name": {
            "type": "string",
            "example": "IT Services"
          },
          "head": {
            "type": "string",
            "example": "pradhyum"
          },
          "memberCount": {
            "type": "number",
            "example": 15
          }
        },
        "required": [
          "id",
          "name",
          "head",
          "memberCount"
        ]
      },
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "D101"
          },
          "name": {
            "type": "string",
            "example": "IT Services"
          },
          "head": {
            "type": "string",
            "example": "Pradhyum"
          },
          "memberCount": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "name",
          "head",
          "memberCount"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "D101"
          },
          "name": {
            "type": "string",
            "example": "IT Services"
          },
          "head": {
            "type": "string",
            "example": "Pradhyum"
          },
          "memberCount": {
            "type": "number",
            "example": 12
          }
        }
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}